+++ resume/models.py
@@ -8,7 +8,7 @@ RESUME_TEMPLATE_CHOICES = (('Simple Resume', 'Simple Resume'), ('Modern Resume',
 
 class Resume(models.Model):
     # Profile owner
-    owner = models.ForeignKey(Profile, on_delete=models.CASCADE)
+    owner = models.ForeignKey(Profile, on_delete=models.CASCADE, blank=True, null=True)
     # Select template
     resume_template = models.CharField(choices=RESUME_TEMPLATE_CHOICES, default='_', blank=True, null=True)
     # General info
@@ -25,7 +25,7 @@ class Resume(models.Model):
 
 
 class Experience(models.Model):
-    owner = models.ForeignKey(Profile, on_delete=models.CASCADE)
+    owner = models.ForeignKey(Profile, on_delete=models.CASCADE, blank=True, null=True)
     resume = models.ForeignKey(Resume, on_delete=models.CASCADE)
     employer_name = models.CharField(max_length=250, blank=True, null=True)
     position = models.CharField(max_length=250, blank=True, null=True)
@@ -40,7 +40,7 @@ class Experience(models.Model):
 
 
 class Projects(models.Model):
-    owner = models.ForeignKey(Profile, on_delete=models.CASCADE)
+    owner = models.ForeignKey(Profile, on_delete=models.CASCADE, blank=True, null=True)
     resume = models.ForeignKey(Resume, on_delete=models.CASCADE)
     project_name = models.CharField(max_length=250, blank=True, null=True)
     start_date_of_project = models.CharField(blank=True, null=True)
@@ -54,7 +54,7 @@ class Projects(models.Model):
 
 
 class Education(models.Model):
-    owner = models.ForeignKey(Profile, on_delete=models.CASCADE)
+    owner = models.ForeignKey(Profile, on_delete=models.CASCADE, blank=True, null=True)
     resume = models.ForeignKey(Resume, on_delete=models.CASCADE)
     school_name = models.CharField(max_length=250, blank=True, null=True)
     degree = models.CharField(max_length=250, blank=True, null=True)
@@ -68,7 +68,7 @@ class Education(models.Model):
 
 
 class Skills(models.Model):
-    owner = models.ForeignKey(Profile, on_delete=models.CASCADE)
+    owner = models.ForeignKey(Profile, on_delete=models.CASCADE, blank=True, null=True)
     resume = models.ForeignKey(Resume, on_delete=models.CASCADE)
     skill_name = models.CharField(max_length=250, blank=True, null=True)
     skill_description = models.TextField(max_length=1000, blank=True, null=True)