+++ resume_builder/urls.py
@@ -5,7 +5,7 @@ from django.conf import settings
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('resume', include('resume.urls')),
+    path('', include('resume.urls')),
     path('', include('users.urls'))
 ]